+++ src/ScottPlot5/ScottPlot5/Interfaces/IAxis.cs
@@ -37,10 +37,10 @@ public interface IAxis : IPanel
     /// <summary>
     /// Given a distance in pixel space, converts to coordinate space
     /// </summary>
-    /// <param name="coordinateDistance">A distance in pixel units</param>
+    /// <param name="pixelDistance">A distance in pixel units</param>
     /// <param name="dataArea">The rectangle onto which the coordinates are mapped</param>
     /// <returns>The same distance in coordinate units</returns>
-    double GetCoordinateDistance(double pixelDistance, PixelRect dataArea);
+    double GetCoordinateDistance(float pixelDistance, PixelRect dataArea);
 
     /// <summary>
     /// Logic for determining tick positions and formatting tick labels
